refactor(scripts/install-template): avoid installation, assert state
