+++ .github/actions/ci/action.yml
@@ -13,12 +13,12 @@ runs:
 
     - name: Get yarn cache
       id: yarn-cache
-      run: echo "::set-output name=dir::$(yarn config get cacheFolder)"
+      run: echo "DIR=$(yarn config get cacheFolder)" >> $GITHUB_OUTPUT
       shell: bash
 
     - uses: actions/cache@v3
       with:
-        path: ${{ steps.yarn-cache.outputs.dir }}
+        path: ${{ steps.yarn-cache.outputs.DIR }}
         key: ${{ runner.os }}-yarn-v1-${{ hashFiles('**/yarn.lock') }}
         restore-keys: |
           ${{ runner.os }}-yarn-v1-
