+++ .github/workflows/release.yml
@@ -35,7 +35,7 @@ jobs:
 
       - name: Storing release version for changeset
         id: release_version
-        run: echo "::set-output name=value::$(./scripts/print_release_version.sh)"
+        run: echo "VALUE=$(./scripts/print_release_version.sh)" >> $GITHUB_OUTPUT
         shell: bash
         env:
           GITHUB_TOKEN: ${{ secrets.RELEASE_GITHUB_TOKEN }}
@@ -49,8 +49,8 @@ jobs:
           version: yarn changeset:version-and-format
           commit: 'ci(changesets): version packages'
           createGithubReleases: aggregate
-          githubReleaseName: v${{ steps.release_version.outputs.value }}
-          githubTagName: v${{ steps.release_version.outputs.value }}
+          githubReleaseName: v${{ steps.release_version.outputs.VALUE }}
+          githubTagName: v${{ steps.release_version.outputs.VALUE }}
         env:
           GITHUB_TOKEN: ${{ secrets.RELEASE_GITHUB_TOKEN }}
           SKIP_POSTINSTALL_DEV_SETUP: true
